feat(kzf): namespace switching
The easiest way to do this, it seems, it just to have fzf `become`
kzf (thereby making this script recursive). That allows creating
early detection of certain flags that will prompt the user
for more information, e.g. the namespace, context, resource, etc.

An early version of this program was a Go application which used fzf's
`listen` feature to send messages back and forth. That was a clever
hack but way too complicated. Even now, this implementation
has twice the features and is less than half the SLOC.