ref(kzf): deduplicate keybind command declarations
There was a lot of overlap between these.

First and foremost, --context was promoted to a common option.
Everywhere that kubectl_common_opts was passed, --context was passed
too. That makes sense.

Next, I created a dict of common binding opts and used those for the
keybinds. I chose not to use a function here because I like having the
choice of [@] and [*] in bash, and reading an array from a command is a
pain.

I like bash for some things and not others. Having everything at the top
level is a big thumbs up from me. Trying to write it like a programming
language is a big thumbs down from me. If I wanted to encapsulate, I'd
just write Go instead.