fix(kzf): some flags always active
This is caused due to recent changes in flag population.
Before 0.4(?) only the flags passed in were tracked internally.
This meant that ${FOO-} was a valid substitution, since if a flag
was present, it had a value.

Now, we have the env vars. Some flags may be unset. It requires us to
check ${FOO:-} to be sure that a flag was present AND not empty.