fix(kzf): boolean flags w/ defaults not remembered during --select-*
This was a doozy. There was no good hack here, so I tried to fix the
root cause (with an even worse hack).

The problem is this: right now, there are two types of variables.

1. Options. When a user passes in some value, that is the
	value. --namespace=foo internally is stored as foo
2. Flags. When a user passes in some value, that value is transformed.
	--viddy=true becomes viddy=--viddy (to simplify command generation).
	--viddy=false *unsets* the viddy var (to differentiate it from "no
	value", which is `""`).

Confused yet? Me too! This whole script is a hack. I'm leaning into
it. Au naturale "vibe coding": I don't care what I'm writing so long as
it works (but, to be clear, *I* am writing it).

Anyway, to fix the issue, I had to represent the above disambiguation in
the code. I couldn't (or didn't want to, or wasn't skilled enough)
mingle those two use cases under one $flags var, so I just split it. Now
there's flags and options and we can deal with the consequences of that
later.

Another fun piece of this is that I now store the default options/flags
separately. Why? Seemed like the easiest way to catalogue a list of
options that this script offers. Seeing as they might be unset.

You never know when a default value might be useful too, though.